alert create+edit+remove, show history graph on sensor with sensor readings

+++ app/controllers/groups_controller.rb
@@ -6,18 +6,25 @@ class GroupsController < ApplicationController
 			redirect_to login_path
 			return
 		end
+		@isAdmin= User.find(session[:user_id]).role
+		@sensorsCount= Array.new
 		id=session[:user_id]
 		@groups=Group.where(user_id:id)#id=1 provisorio
+		@groups.each do |g|
+			@sensorsCount[g.id]=Sensor.where(groups_id: g.id).count()
+		end
+
 		render "groups_controller/index"
 	end
 
 	def new
 		#criar grupo POST
 		if session[:user_id]==nil
- 			redirect_to login_path
- 			return
- 		end
- 		id=session[:user_id]
+			redirect_to login_path
+			return
+		end
+		@isAdmin= User.find(session[:user_id]).role
+		id=session[:user_id]
 		group = Group.new
 		group.user_id=id; #provisorio
 		group.description=params[:desc]
@@ -27,18 +34,20 @@ class GroupsController < ApplicationController
 
 	def edit
 		if session[:user_id]==nil
- 			redirect_to login_path
- 			return
- 		end
+			redirect_to login_path
+			return
+		end
+		@isAdmin= User.find(session[:user_id]).role
 		@group = Group.find(params[:id])
 		render "groups_controller/editGroup"
 	end
 
 	def editPost
 		if session[:user_id]==nil
- 			redirect_to login_path
- 			return
- 		end
+			redirect_to login_path
+			return
+		end
+		@isAdmin= User.find(session[:user_id]).role
 		@group= Group.find(params[:id])
 		@group.description= params[:desc]
 		@group.save
@@ -47,11 +56,15 @@ class GroupsController < ApplicationController
 
 	def delete
 		if session[:user_id]==nil
- 			redirect_to login_path
- 			return
- 		end
+			redirect_to login_path
+			return
+		end
+		@isAdmin= User.find(session[:user_id]).role
 		g = Group.find(params[:id])
-		g.destroy
+		sensorsCount=Sensor.where(groups_id: g.id).count()
+		if sensorsCount==0
+			g.destroy
+		end
 		redirect_to mygroups_path
 	end
